# LocalAIAgent.API/openapi/LocalAIAgent.API.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LocalAIAgent.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/auth/register/options": {
      "post": {
        "tags": [
          "Fido2"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationOptionsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationOptionsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationOptionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialCreateOptions"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialCreateOptions"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialCreateOptions"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register/complete": {
      "post": {
        "tags": [
          "Fido2"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRegistrationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRegistrationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredPublicKeyCredential"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredPublicKeyCredential"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredPublicKeyCredential"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/News/GetExpandedNews": {
      "post": {
        "tags": [
          "News"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExpandedNewsResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpandedNewsResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpandedNewsResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/News/Feedback": {
      "post": {
        "tags": [
          "News"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsFeedbackDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsFeedbackDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewsFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Algorithm": {
        "enum": [
          -65535,
          -259,
          -258,
          -257,
          -47,
          -39,
          -38,
          -37,
          -36,
          -35,
          -8,
          -7
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "AttestationConveyancePreference": {
        "enum": [
          "none",
          "indirect",
          "direct",
          "enterprise"
        ],
        "type": "string"
      },
...
      "AttestationStatementFormatIdentifier": {
        "enum": [
          "packed",
          "tpm",
          "android-key",
          "android-safetynet",
          "fido-u2f",
          "apple",
          "none"
        ],
        "type": "string"
      },
      "AuthenticationExtensionsClientInputs": {
        "type": "object",
        "properties": {
          "example.extension.bool": {
            "type": "boolean",
            "nullable": true
          },
          "exts": {
            "type": "boolean",
            "nullable": true
          },
          "uvm": {
            "type": "boolean",
            "nullable": true,
            "writeOnly": true
          },
          "credProps": {
            "type": "boolean",
            "nullable": true
          },
          "prf": {
            "$ref": "#/components/schemas/AuthenticationExtensionsPRFInputs"
          },
          "largeBlob": {
            "$ref": "#/components/schemas/AuthenticationExtensionsLargeBlobInputs"
          },
          "credentialProtectionPolicy": {
            "$ref": "#/components/schemas/CredentialProtectionPolicy"
          },
          "enforceCredentialProtectionPolicy": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AuthenticationExtensionsLargeBlobInputs": {
        "type": "object",
        "properties": {
          "support": {
            "$ref": "#/components/schemas/LargeBlobSupport"
          },
          "read": {
            "type": "boolean"
          },
          "write": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AuthenticationExtensionsPRFInputs": {
        "type": "object",
        "properties": {
          "eval": {
            "$ref": "#/components/schemas/AuthenticationExtensionsPRFValues"
          },
          "evalByCredential": {
            "$ref": "#/components/schemas/StringAuthenticationExtensionsPRFValuesKeyValuePair"
          }
        },
        "additionalProperties": false
      },
...
      "AuthenticatorAttachment": {
        "enum": [
          "platform",
          "cross-platform"
        ],
        "type": "string"
      },
      "AuthenticatorAttestationRawResponse": {
        "required": [
          "clientExtensionResults",
          "id",
          "rawId",
          "response",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "rawId": {
            "type": "string",
            "format": "byte"
          },
          "type": {
            "$ref": "#/components/schemas/PublicKeyCredentialType"
          },
          "response": {
            "$ref": "#/components/schemas/AttestationResponse"
          },
          "extensions": {
            "$ref": "#/components/schemas/AuthenticationExtensionsClientOutputs"
          },
          "clientExtensionResults": {
            "$ref": "#/components/schemas/AuthenticationExtensionsClientOutputs"
          }
        },
        "additionalProperties": false
      },
      "AuthenticatorSelection": {
        "type": "object",
        "properties": {
          "authenticatorAttachment": {
            "$ref": "#/components/schemas/AuthenticatorAttachment"
          },
          "residentKey": {
            "$ref": "#/components/schemas/ResidentKeyRequirement"
          },
          "requireResidentKey": {
            "type": "boolean",
            "deprecated": true
          },
          "userVerification": {
            "$ref": "#/components/schemas/UserVerificationRequirement"
          }
        },
        "additionalProperties": false
      },
      "AuthenticatorTransport": {
        "enum": [
          "usb",
          "nfc",
          "ble",
          "smart-card",
          "hybrid",
          "internal"
        ],
        "type": "string"
      },
...
      "CredentialCreateOptions": {
        "required": [
          "challenge",
          "pubKeyCredParams",
          "rp",
          "user"
        ],
        "type": "object",
        "properties": {
          "rp": {
            "$ref": "#/components/schemas/PublicKeyCredentialRpEntity"
          },
          "user": {
            "$ref": "#/components/schemas/Fido2User"
          },
          "challenge": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "pubKeyCredParams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PubKeyCredParam"
            },
            "nullable": true
          },
          "timeout": {
            "type": "integer",
            "format": "int64"
          },
          "attestation": {
            "$ref": "#/components/schemas/AttestationConveyancePreference"
          },
          "attestationFormats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttestationStatementFormatIdentifier"
            },
            "nullable": true
          },
          "authenticatorSelection": {
            "$ref": "#/components/schemas/AuthenticatorSelection"
          },
          "hints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicKeyCredentialHint"
            },
            "nullable": true
          },
          "excludeCredentials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicKeyCredentialDescriptor"
            },
            "nullable": true
          },
          "extensions": {
            "$ref": "#/components/schemas/AuthenticationExtensionsClientInputs"
          }
        },
        "additionalProperties": false
      },
...
      "CredentialProtectionPolicy": {
        "enum": [
          "userVerificationOptional",
          "userVerificationOptionalWithCredentialIDList",
          "userVerificationRequired"
        ],
        "type": "string"
      },
      "CredentialRegistrationRequest": {
        "required": [
          "attestation",
          "credentialName"
        ],
        "type": "object",
        "properties": {
          "attestation": {
            "$ref": "#/components/schemas/AuthenticatorAttestationRawResponse"
          },
          "credentialName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExpandedNewsResult": {
        "type": "object",
        "properties": {
          "articleWasTranslated": {
            "type": "boolean"
          },
          "translation": {
            "type": "string",
            "nullable": true
          },
          "termsAndExplanations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TermStringExplanationStringKeyValuePair"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Fido2User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NewsFeedbackDto": {
        "required": [
          "articleLink",
          "articleSummary",
          "articleTitle",
          "articleTopic",
          "isLiked"
        ],
        "type": "object",
        "properties": {
          "articleLink": {
            "type": "string",
            "nullable": true
          },
          "articleTitle": {
            "type": "string",
            "nullable": true
          },
          "articleSummary": {
            "type": "string",
            "nullable": true
          },
          "articleTopic": {
            "type": "string",
            "nullable": true
          },
          "isLiked": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "selectedLikes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "selectedDislikes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PubKeyCredParam": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/PublicKeyCredentialType"
          },
          "alg": {
            "$ref": "#/components/schemas/Algorithm"
          }
        },
        "additionalProperties": false
      },
      "PublicKeyCredentialDescriptor": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/PublicKeyCredentialType"
          },
          "id": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "transports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthenticatorTransport"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PublicKeyCredentialHint": {
        "enum": [
          "security-key",
          "client-device",
          "hybrid"
        ],
        "type": "string"
      },
      "PublicKeyCredentialRpEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PublicKeyCredentialType": {
        "enum": [
          "public-key",
          "invalid"
        ],
        "type": "string"
      },
      "RegisteredPublicKeyCredential": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/PublicKeyCredentialType"
          },
          "id": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "publicKey": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "transports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthenticatorTransport"
            },
            "nullable": true
          },
          "signCount": {
            "type": "integer",
            "format": "int32"
          },
          "isBackupEligible": {
            "type": "boolean"
          },
          "isBackedUp": {
            "type": "boolean"
          },
          "aaGuid": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/Fido2User"
          },
          "attestationFormat": {
            "type": "string",
            "nullable": true
          },
          "attestationObject": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "attestationClientDataJson": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegistrationOptionsRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "inviteToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ResidentKeyRequirement": {
        "enum": [
          "required",
          "preferred",
          "discouraged"
        ],
        "type": "string"
      },
...
      "TermStringExplanationStringKeyValuePair": {
        "type": "object",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/TermString"
          },
          "value": {
            "$ref": "#/components/schemas/ExplanationString"
          }
        },
        "additionalProperties": false
      },
...
      "UserVerificationRequirement": {
        "enum": [
          "required",
          "preferred",
          "discouraged"
        ],
        "type": "string"
      },
...
    }
  },
...
}